Update project with failing tests
